+++ src/provider/locationiq/LocationIQGeocodeQuery.ts
@@ -1,13 +1,13 @@
 import { GeocodeQuery, GeocodeQueryObject } from "query";
 
-export interface NominatimGeocodeQueryObject extends GeocodeQueryObject {
+export interface LocationIQGeocodeQueryObject extends GeocodeQueryObject {
   readonly countryCodes?: string[];
   readonly excludePlaceIds?: number[];
   readonly bounded?: boolean;
   readonly dedupe?: boolean;
 }
 
-export default class NominatimGeocodeQuery extends GeocodeQuery {
+export default class LocationIQGeocodeQuery extends GeocodeQuery {
   private readonly countryCodes?: string[];
 
   private readonly excludePlaceIds?: number[];
@@ -23,7 +23,7 @@ export default class NominatimGeocodeQuery extends GeocodeQuery {
     bounds,
     dedupe,
     ...geocodeQueryObject
-  }: NominatimGeocodeQueryObject) {
+  }: LocationIQGeocodeQueryObject) {
     super({ bounds, ...geocodeQueryObject });
     this.countryCodes = countryCodes;
     this.excludePlaceIds = excludePlaceIds;
@@ -37,12 +37,12 @@ export default class NominatimGeocodeQuery extends GeocodeQuery {
   }
 
   public static create(
-    object: NominatimGeocodeQueryObject
-  ): NominatimGeocodeQuery {
+    object: LocationIQGeocodeQueryObject
+  ): LocationIQGeocodeQuery {
     return new this(object);
   }
 
-  public toObject(): NominatimGeocodeQueryObject {
+  public toObject(): LocationIQGeocodeQueryObject {
     return {
       ...super.toObject(),
       countryCodes: this.countryCodes,
@@ -52,32 +52,34 @@ export default class NominatimGeocodeQuery extends GeocodeQuery {
     };
   }
 
-  public withCountryCodes(countryCodes: string[]): NominatimGeocodeQuery {
-    return new NominatimGeocodeQuery({ ...this.toObject(), countryCodes });
+  public withCountryCodes(countryCodes: string[]): LocationIQGeocodeQuery {
+    return new LocationIQGeocodeQuery({ ...this.toObject(), countryCodes });
   }
 
   public getCountryCodes(): undefined | string[] {
     return this.countryCodes;
   }
 
-  public withExcludePlaceIds(excludePlaceIds: number[]): NominatimGeocodeQuery {
-    return new NominatimGeocodeQuery({ ...this.toObject(), excludePlaceIds });
+  public withExcludePlaceIds(
+    excludePlaceIds: number[]
+  ): LocationIQGeocodeQuery {
+    return new LocationIQGeocodeQuery({ ...this.toObject(), excludePlaceIds });
   }
 
   public getExcludePlaceIds(): undefined | number[] {
     return this.excludePlaceIds;
   }
 
-  public withBounded(bounded: boolean): NominatimGeocodeQuery {
-    return new NominatimGeocodeQuery({ ...this.toObject(), bounded });
+  public withBounded(bounded: boolean): LocationIQGeocodeQuery {
+    return new LocationIQGeocodeQuery({ ...this.toObject(), bounded });
   }
 
   public getBounded(): undefined | boolean {
     return this.bounded;
   }
 
-  public withDedupe(dedupe: boolean): NominatimGeocodeQuery {
-    return new NominatimGeocodeQuery({ ...this.toObject(), dedupe });
+  public withDedupe(dedupe: boolean): LocationIQGeocodeQuery {
+    return new LocationIQGeocodeQuery({ ...this.toObject(), dedupe });
   }
 
   public getDedupe(): undefined | boolean {